+++ app/src/main/java/com/eyeson/android/MainActivity.kt
@@ -19,8 +19,6 @@ class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.main_activity)
-
         setContent {
             val systemUiController = rememberSystemUiController()
             SideEffect {
@@ -48,6 +46,7 @@ class MainActivity : AppCompatActivity() {
         }
         return
         // view world
+        setContentView(R.layout.main_activity)
         if (savedInstanceState == null) {
             supportFragmentManager.beginTransaction()
                 .replace(R.id.container, ConnectionFragment.newInstance())